win: Fix implicit conversion error
Fixes the below error when building with use_custom_libcxx=false.

error: implicit conversion changes signedness: 'size_t' (aka 'unsigned int') to 'ptrdiff_t' (aka 'int') [-Werror,-Wsign-conversion]
                       std::begin(battery_reporting_scales) + nb_elements,
                                                            ~ ^~~~~~~~~~~

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219163
Commit-Queue: Marshall Greenblatt <marshall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101169}

+++ base/power_monitor/battery_level_provider_win.cc
@@ -18,6 +18,7 @@
 
 #include "base/memory/weak_ptr.h"
 #include "base/metrics/histogram_macros.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/task/sequenced_task_runner.h"
 #include "base/task/task_traits.h"
 #include "base/task/thread_pool.h"
@@ -134,7 +135,8 @@ absl::optional<uint32_t> GetBatteryBatteryDischargeGranularity(
   if (!success)
     return absl::nullopt;
 
-  size_t nb_elements = bytes_returned / sizeof(BATTERY_REPORTING_SCALE);
+  ptrdiff_t nb_elements = base::checked_cast<ptrdiff_t>(
+      bytes_returned / sizeof(BATTERY_REPORTING_SCALE));
   if (!nb_elements)
     return absl::nullopt;
